Rework ColumnData
It's now called "Vector" and provides information and typed data over a duckdb
vector. The main difference is that the type information that used to be read
on each data_chunk, is now read upfront and that information is re-used for each
chunk. This should yield a bit better performance, and also makes it easier to
manage memory. Specifically, all logical_types should now be properly cleaned up.